requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langgraph>=0.0.20
python-dotenv>=1.0.0
jupyter>=1.0.0
matplotlib>=3.7.0
pandas>=2.0.0
